+++ lib/core/bloc_logger_observer.dart
@@ -6,20 +6,21 @@ import 'log_storage_service.dart';
 /// A custom [BlocObserver] that logs BLoC events, state changes, and errors
 /// to the `simple_logger_overlay` system.
 ///
-/// This observer writes logs into persistent storage using [LogStorageService],
-/// and formats them as [SimpleLog] entries for display in the overlay.
+/// This observer writes logs into persistent storage using [SimpleOverlayLogStorageService],
+/// and formats them as [SimpleOverlayLog] entries for display in the overlay.
 ///
 /// Usage:
 /// ```dart
 /// Bloc.observer = SimpleOverlayBlocObserverLogger();
 /// ```
 class SimpleOverlayBlocObserverLogger extends BlocObserver {
-  final LogStorageService _storage = LogStorageService();
+  final SimpleOverlayLogStorageService _storage =
+      SimpleOverlayLogStorageService();
 
-  /// Logs every dispatched BLoC [event] as a `debug`-level [SimpleLog].
+  /// Logs every dispatched BLoC [event] as a `debug`-level [SimpleOverlayLog].
   @override
   void onEvent(Bloc bloc, Object? event) {
-    _storage.addSimpleLog(SimpleLog(
+    _storage.addSimpleLog(SimpleOverlayLog(
       timestamp: DateTime.now(),
       tag: bloc.runtimeType.toString(),
       level: LogLevel.debug,
@@ -28,10 +29,10 @@ class SimpleOverlayBlocObserverLogger extends BlocObserver {
     super.onEvent(bloc, event);
   }
 
-  /// Logs every state [change] as an `info`-level [SimpleLog].
+  /// Logs every state [change] as an `info`-level [SimpleOverlayLog].
   @override
   void onChange(BlocBase bloc, Change change) {
-    _storage.addSimpleLog(SimpleLog(
+    _storage.addSimpleLog(SimpleOverlayLog(
       timestamp: DateTime.now(),
       tag: bloc.runtimeType.toString(),
       level: LogLevel.info,
@@ -40,10 +41,10 @@ class SimpleOverlayBlocObserverLogger extends BlocObserver {
     super.onChange(bloc, change);
   }
 
-  /// Logs all [error]s and [stackTrace]s as `error`-level [SimpleLog]s.
+  /// Logs all [error]s and [stackTrace]s as `error`-level [SimpleOverlayLog]s.
   @override
   void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
-    _storage.addSimpleLog(SimpleLog(
+    _storage.addSimpleLog(SimpleOverlayLog(
       timestamp: DateTime.now(),
       tag: bloc.runtimeType.toString(),
       level: LogLevel.error,