+++ lib/core/logger_controller.dart
@@ -4,13 +4,14 @@ import 'package:shake/shake.dart';
 import '../simple_logger_overlay.dart';
 
 /// Controls the logger overlay, handles logging, and manages storage.
-class LoggerController {
-  static final LoggerController _instance = LoggerController._internal();
+class SimpleOverlayLoggerShakeController {
+  static final SimpleOverlayLoggerShakeController _instance =
+      SimpleOverlayLoggerShakeController._internal();
   ShakeDetector? _shakeDetector;
 
-  LoggerController._internal();
+  SimpleOverlayLoggerShakeController._internal();
 
-  factory LoggerController() => _instance;
+  factory SimpleOverlayLoggerShakeController() => _instance;
 
   /// Initializes the overlay and logger options.
   ///
@@ -18,12 +19,16 @@ class LoggerController {
   /// [enableShake] enables shake-to-open gesture in debug mode.
   /// [enableInRelease] allows logger in release mode (disabled by default).
   /// [performanceMode] disables file writes for high-FPS apps (e.g., games).
-  void init({required BuildContext context, bool enableShake = true}) {
+  void init(
+      {required BuildContext context,
+      bool enableShake = true,
+      GlobalKey<NavigatorState>? navigatorKey}) {
     _shakeDetector?.stopListening(); // clean up old
 
     if (enableShake) {
       _shakeDetector = ShakeDetector.autoStart(
-        onPhoneShake: (event) => SimpleLoggerOverlay.show(context),
+        onPhoneShake: (event) =>
+            SimpleLoggerOverlay.show(context, navigatorKey: navigatorKey),
         shakeThresholdGravity: 2.7,
       );
     }