+++ lib/core/riverpod_logger.dart
@@ -6,15 +6,16 @@ import 'log_storage_service.dart';
 /// A custom [ProviderObserver] that logs Riverpod provider changes
 /// to the `simple_logger_overlay` system.
 ///
-/// This observer writes logs into persistent storage using [LogStorageService],
-/// and formats them as [SimpleLog] entries for display in the overlay.
+/// This observer writes logs into persistent storage using [SimpleOverlayLogStorageService],
+/// and formats them as [SimpleOverlayLog] entries for display in the overlay.
 class SimpleOverlayLoggerRiverpodObserver extends ProviderObserver {
-  final LogStorageService _storage = LogStorageService();
+  final SimpleOverlayLogStorageService _storage =
+      SimpleOverlayLogStorageService();
 
   @override
   void didUpdateProvider(ProviderBase provider, Object? previousValue,
       Object? newValue, ProviderContainer container) {
-    _storage.addSimpleLog(SimpleLog(
+    _storage.addSimpleLog(SimpleOverlayLog(
       timestamp: DateTime.now(),
       tag: provider.name ?? provider.runtimeType.toString(),
       level: LogLevel.debug,
@@ -24,7 +25,7 @@ class SimpleOverlayLoggerRiverpodObserver extends ProviderObserver {
 
   @override
   void didDisposeProvider(ProviderBase provider, ProviderContainer container) {
-    _storage.addSimpleLog(SimpleLog(
+    _storage.addSimpleLog(SimpleOverlayLog(
       timestamp: DateTime.now(),
       tag: provider.name ?? provider.runtimeType.toString(),
       level: LogLevel.info,